+++ opencv/opencv-closeups/closeup.py
@@ -9,8 +9,11 @@ import tempfile
 import subprocess
 import os
 import itertools
+import numpy as np
+import cv2
 import time
 from config import *
+
 
 def main():
     global logger
@@ -59,63 +62,79 @@ def main():
     connection.close()
  
 
-def ocr(filename, tmpfilename):
-    text=""
-    tmpfile=None
+
+def detect_closeup(inputfile, ext, host, fileid, key):
+    global logger, extractorName
+    global sslVerify, face_cascade_path, profileface_cascade_path
+
+    logger.debug("INSIDE: detect_closeup")
     try:
-        subprocess.check_call(["tesseract", filename, tmpfilename])
-        tmpfile="./"+tmpfilename+".txt"
-        with open(tmpfile, 'r') as f:
-            text = f.read()
+        face_cascade = cv2.CascadeClassifier(face_cascade_path)
+        profile_face_cascade = cv2.CascadeClassifier(profileface_cascade_path)
+               
+        img = cv2.imread(inputfile, cv2.CV_LOAD_IMAGE_GRAYSCALE)
+        if img is not None:
+            gray=img
+            gray = cv2.equalizeHist(gray)
+            
+            imgh=len(gray)
+            imgw=len(gray[0])
+        
+            midCloseUp=False
+            fullCloseUp=False
+
+            faces=face_cascade.detectMultiScale(gray, scaleFactor=1.2, minNeighbors=2, minSize=(imgw/8, imgh/8), flags=cv2.cv.CV_HAAR_SCALE_IMAGE)
+            for (x,y,w,h) in faces:
+                if ((w*h>=(imgw*imgh/3)) or (w>=0.8*imgw and h>=0.5*imgh) or (w>=0.5*imgw and h>=0.8*imgh)): #this is a closeup
+                    fullCloseUp=True
+            for (x,y,w,h) in faces: 
+                if(w*h>=(imgw*imgh/8)): #this is a medium closeup
+                    midCloseUp=True
+            
+            profilefaces=profile_face_cascade.detectMultiScale(gray, scaleFactor=1.05, minNeighbors=10, minSize=(imgw/8, imgh/8))
+            for (x,y,w,h) in profilefaces:
+                if ((w*h>=(imgw*imgh/3)) or (w>=0.8*imgw and h>=0.5*imgh) or (w>=0.5*imgw and h>=0.8*imgh)): #this is a closeup
+                    fullCloseUp=True
+            for (x,y,w,h) in profilefaces: 
+                if(w*h>=(imgw*imgh/8)): #this is a medium closeup
+                    midCloseUp=True
+
+
+            headers={'Content-Type': 'application/json'}
+            if fullCloseUp:
+                url=host+'api/files/'+ fileid +'/tags?key=' + key
+                mdata={}
+                mdata["tags"]=["Full Close Up Automatically Detected"]
+                mdata["extractor_id"]=extractorName
+                logger.debug("tags: %s",json.dumps(mdata))
+                rt = requests.post(url, headers=headers, data=json.dumps(mdata), verify=sslVerify)
+                rt.raise_for_status()
+                logger.debug("[%s] created section and previews of type %s", fileid, ext)
+
+            elif midCloseUp:
+                url=host+'api/files/'+ fileid +'/tags?key=' + key
+                mdata={}
+                mdata["tags"]=["Mid Close Up Automatically Detected"]
+                mdata["extractor_id"]=extractorName
+                logger.debug("tags: %s",json.dumps(mdata))
+                rt = requests.post(url, headers=headers, data=json.dumps(mdata), verify=sslVerify)
+                rt.raise_for_status()
+                logger.debug("[%s] created section and previews of type %s", fileid, ext)        
     finally:
-        if tmpfile is not None and os.path.isfile(tmpfile):
-            os.remove(tmpfile)
-        return clean_text(text)
-
-def clean_text(text):
-    t = ""
-    words=text.split()
-    for word in words:
-        w = clean_word(word) 
-        if w != "":
-            t+= w + " "
-    return t
-
-def clean_word(word):
-    cw = word.strip('(){}[].,')
-    if cw.isalnum() and len(cw)>=2:
-        return cw
-    else:
-        return ""
-
-
-def extract_OCR(inputfile, host, fileid, key):
-    global logger, sslVerify
-    global extractorName
-    logger.debug("INSIDE: extract_OCR")
-    
-    try:
-        ocrtext = ocr(inputfile, "tmpocr")
-
-        headers={'Content-Type': 'application/json'}
-
-        url=host+'api/files/'+ fileid +'/metadata?key=' + key
-        mdata={}
-        mdata["extractor_id"]=extractorName
-        mdata["ocr_simple"]=[ocrtext]
-
-        logger.debug("metadata: %s",json.dumps(mdata))
-        rt = requests.post(url, headers=headers, data=json.dumps(mdata), verify=sslVerify)
-        rt.raise_for_status()
-        logger.debug("[%s] simple ocr performed successfully", fileid)
-
-    finally:
-        logger.debug("[%s] done with simple ocr extractor", fileid)  
+        #os.remove(previewfile)     
+        logger.debug("done with closeups")  
         
 
+def get_image_data(imagefile):
+    global logger
+
+    text=subprocess.check_output(['identify', imagefile], stderr=subprocess.STDOUT)
+    return text
+
 def on_message(channel, method, header, body):
-    global logger, sslVerify
-    global extractorName
+    global logger, extractorName
+    global sslVerify
+    
     statusreport = {}
 
     inputfile=None
@@ -152,7 +171,8 @@ def on_message(channel, method, header, body):
             for chunk in r.iter_content(chunk_size=10*1024):
                 f.write(chunk)
 
-        statusreport['status'] = 'OCRing image and associating text with file.'
+        
+        statusreport['status'] = 'Detecting closeup and tagging.'
         statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S')
         statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S')
         channel.basic_publish(exchange='',
@@ -162,15 +182,15 @@ def on_message(channel, method, header, body):
                             body=json.dumps(statusreport))
 
 
-
-        extract_OCR(inputfile, host, fileid, key)
+        # create previews
         
-
+        detect_closeup(inputfile, 'jpg', host, fileid, key)
+               
     except subprocess.CalledProcessError as e:
         logger.exception("[%s] error processing [exit code=%d]\n%s", fileid, e.returncode, e.output)
         statusreport['status'] = 'Error processing.'
-        statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S')
-        statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S') 
+        statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S') 
+        statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S')
         channel.basic_publish(exchange='',
                 routing_key=header.reply_to,
                 properties=pika.BasicProperties(correlation_id = \
@@ -179,8 +199,8 @@ def on_message(channel, method, header, body):
     except:
         logger.exception("[%s] error processing", fileid)
         statusreport['status'] = 'Error processing.'
-        statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S')
-        statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S') 
+        statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S') 
+        statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S')
         channel.basic_publish(exchange='',
                 routing_key=header.reply_to,
                 properties=pika.BasicProperties(correlation_id = \
@@ -195,9 +215,12 @@ def on_message(channel, method, header, body):
                             properties=pika.BasicProperties(correlation_id = \
                                                         header.correlation_id),
                             body=json.dumps(statusreport))
-        if inputfile is not None:
-            os.remove(inputfile)
-            
+        if inputfile is not None and os.path.isfile(inputfile):
+            try:
+                os.remove(inputfile)
+            except OSError as oserror:
+                logger.exception("[%s] error removing input file: \n %s", fileid, oserror)
+
         # Ack
         channel.basic_ack(method.delivery_tag)
         logger.debug("[%s] finished processing", fileid)
