+++ opencv/opencv-profiles/profile.py
@@ -9,6 +9,8 @@ import tempfile
 import subprocess
 import os
 import itertools
+import numpy as np
+import cv2
 import time
 from config import *
 
@@ -59,63 +61,107 @@ def main():
     connection.close()
  
 
-def ocr(filename, tmpfilename):
-    text=""
-    tmpfile=None
+
+def create_image_section(inputfile, ext, host, fileid, key):
+    global logger, extractorName
+    global sslVerify, profileface_cascade_path
+
+    logger.debug("INSIDE: create_image_section")
+
+    sectionfile=None
+
     try:
-        subprocess.check_call(["tesseract", filename, tmpfilename])
-        tmpfile="./"+tmpfilename+".txt"
-        with open(tmpfile, 'r') as f:
-            text = f.read()
+
+        profile_face_cascade = cv2.CascadeClassifier(profileface_cascade_path)
+
+        img = cv2.imread(inputfile, cv2.CV_LOAD_IMAGE_GRAYSCALE)
+        
+        #gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+        if img is not None:
+            gray = img
+            gray = cv2.equalizeHist(gray)
+
+            faces=profile_face_cascade.detectMultiScale(gray, scaleFactor=1.05, minNeighbors=10)
+            for (x,y,w,h) in faces:
+                roi_color = img[y:y+h, x:x+w]
+                (fd, sectionfile)=tempfile.mkstemp(suffix='.' + ext)
+                os.close(fd)
+                cv2.imwrite(sectionfile, roi_color)
+
+                url=host + 'api/sections?key=' + key
+                logger.debug("url=%s",url)
+                secdata={}
+                secdata["file_id"]=fileid
+                #print(type(fileid),type(x),type(y),type(w),type(h))
+                secdata["area"]={"x":int(x), "y":int(y),"w":int(w),"h":int(h)}
+                
+                #logger.debug("section json [%s]",(json.dumps(secdata)))
+                
+                headers={'Content-Type': 'application/json'}
+               
+                r = requests.post(url,headers=headers, data=json.dumps(secdata), verify=sslVerify)
+                r.raise_for_status()
+                
+                sectionid=r.json()['id']
+                logger.debug(("section id [%s]",sectionid))
+
+                url=host + 'api/previews?key=' + key
+                # upload preview image
+                with open(sectionfile, 'rb') as f:
+                    files={"File" : f}
+                    rc = requests.post(url, files=files, verify=sslVerify)
+                    rc.raise_for_status()
+                previewid = rc.json()['id']
+                logger.debug("preview id=[%s]",rc.json()['id'])
+
+                # associate uploaded image with section
+                imgdata={}
+                imgdata['section_id']=sectionid
+                imgdata['width']=str(w)
+                imgdata['height']=str(h)
+                imgdata['extractor_id']=extractorName
+            
+                headers={'Content-Type': 'application/json'}
+                url = host + 'api/previews/' + previewid + '/metadata?key=' + key
+                # url=host + 'api/files/' + fileid + '/previews/' + previewid + '?key=' + key
+                rp = requests.post(url, headers=headers, data=json.dumps(imgdata), verify=sslVerify);
+                rp.raise_for_status()
+                
+
+                url=host+'api/sections/'+ sectionid+'/tags?key=' + key
+                mdata={}
+                mdata["tags"]=["Human Profile Automatically Detected"]
+                mdata["extractor_id"]=extractorName
+                logger.debug("tags: %s",json.dumps(mdata))
+                rt = requests.post(url, headers=headers, data=json.dumps(mdata), verify=sslVerify)
+                rt.raise_for_status()
+                logger.debug("[%s] created section and previews of type %s", fileid, ext)
+            if len(faces)>=1:
+                url=host+'api/files/'+ fileid+'/tags?key=' + key
+                mdata={}
+                mdata["tags"]=["Human Profile Automatically Detected"]
+                mdata["extractor_id"]=extractorName
+                logger.debug("tags: %s",json.dumps(mdata))
+                rtf = requests.post(url, headers=headers, data=json.dumps(mdata), verify=sslVerify)
+                rtf.raise_for_status()
+                logger.debug("[%s] created section and previews of type %s", fileid, ext)
+
     finally:
-        if tmpfile is not None and os.path.isfile(tmpfile):
-            os.remove(tmpfile)
-        return clean_text(text)
-
-def clean_text(text):
-    t = ""
-    words=text.split()
-    for word in words:
-        w = clean_word(word) 
-        if w != "":
-            t+= w + " "
-    return t
-
-def clean_word(word):
-    cw = word.strip('(){}[].,')
-    if cw.isalnum() and len(cw)>=2:
-        return cw
-    else:
-        return ""
-
-
-def extract_OCR(inputfile, host, fileid, key):
-    global logger, sslVerify
-    global extractorName
-    logger.debug("INSIDE: extract_OCR")
-    
-    try:
-        ocrtext = ocr(inputfile, "tmpocr")
-
-        headers={'Content-Type': 'application/json'}
-
-        url=host+'api/files/'+ fileid +'/metadata?key=' + key
-        mdata={}
-        mdata["extractor_id"]=extractorName
-        mdata["ocr_simple"]=[ocrtext]
-
-        logger.debug("metadata: %s",json.dumps(mdata))
-        rt = requests.post(url, headers=headers, data=json.dumps(mdata), verify=sslVerify)
-        rt.raise_for_status()
-        logger.debug("[%s] simple ocr performed successfully", fileid)
-
-    finally:
-        logger.debug("[%s] done with simple ocr extractor", fileid)  
+        if sectionfile is not None and os.path.isfile(sectionfile):     
+            os.remove(sectionfile)     
+       
         
 
+def get_image_data(imagefile):
+    global logger
+
+    text=subprocess.check_output(['identify', imagefile], stderr=subprocess.STDOUT)
+    return text
+
 def on_message(channel, method, header, body):
-    global logger, sslVerify
-    global extractorName
+    global logger, extractorName
+    global sslVerify
+    
     statusreport = {}
 
     inputfile=None
@@ -131,7 +177,7 @@ def on_message(channel, method, header, body):
 
         # print what we are doing
         logger.debug("[%s] started processing", fileid)
-        # for status reports
+         # for status reports
         statusreport['file_id'] = fileid
         statusreport['extractor_id'] = extractorName
         statusreport['status'] = 'Downloading image file.'
@@ -152,7 +198,7 @@ def on_message(channel, method, header, body):
             for chunk in r.iter_content(chunk_size=10*1024):
                 f.write(chunk)
 
-        statusreport['status'] = 'OCRing image and associating text with file.'
+        statusreport['status'] = 'Extracting profile from image and creating a section.'
         statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S')
         statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S')
         channel.basic_publish(exchange='',
@@ -162,15 +208,19 @@ def on_message(channel, method, header, body):
                             body=json.dumps(statusreport))
 
 
-
-        extract_OCR(inputfile, host, fileid, key)
+        # create previews
+        #create_image_preview(inputfile, 'jpg', '800x600>', host, fileid, key)
+        create_image_section(inputfile, 'jpg', host, fileid, key)
+        #create_image_preview(inputfile, 'jpg', '800x600>', host, fileid, key, '-rotate', '90')
+        #create_image_preview(inputfile, 'jpg', '800x600>', host, fileid, key, '-rotate', '180')
+        #create_image_preview(inputfile, 'jpg', '800x600>', host, fileid, key, '-rotate', '270')
         
 
     except subprocess.CalledProcessError as e:
         logger.exception("[%s] error processing [exit code=%d]\n%s", fileid, e.returncode, e.output)
         statusreport['status'] = 'Error processing.'
-        statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S')
-        statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S') 
+        statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S') 
+        statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S')
         channel.basic_publish(exchange='',
                 routing_key=header.reply_to,
                 properties=pika.BasicProperties(correlation_id = \
@@ -179,8 +229,8 @@ def on_message(channel, method, header, body):
     except:
         logger.exception("[%s] error processing", fileid)
         statusreport['status'] = 'Error processing.'
-        statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S')
-        statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S') 
+        statusreport['start'] = time.strftime('%Y-%m-%dT%H:%M:%S') 
+        statusreport['end']=time.strftime('%Y-%m-%dT%H:%M:%S')
         channel.basic_publish(exchange='',
                 routing_key=header.reply_to,
                 properties=pika.BasicProperties(correlation_id = \
@@ -195,8 +245,12 @@ def on_message(channel, method, header, body):
                             properties=pika.BasicProperties(correlation_id = \
                                                         header.correlation_id),
                             body=json.dumps(statusreport))
-        if inputfile is not None:
-            os.remove(inputfile)
+
+        if inputfile is not None and os.path.isfile(inputfile):
+            try:
+                os.remove(inputfile)
+            except OSError as oserror:
+                logger.exception("[%s] error removing input file: \n %s", fileid, oserror)
             
         # Ack
         channel.basic_ack(method.delivery_tag)
